+++ test_soc/nexys_video/tcl/create_axi_converter.tcl
@@ -16,7 +16,7 @@ if { [get_ips -quiet clk_converter] eq "" } {
 set_property -dict [list \
   CONFIG.DATA_WIDTH $bus_width \
   CONFIG.ID_WIDTH {4} \
-  CONFIG.ADDR_WIDTH {30} ] [get_ips clk_converter]
+  CONFIG.ADDR_WIDTH {31} ] [get_ips clk_converter]
 generate_target {instantiation_template} [get_ips clk_converter]
 create_ip_run [get_ips clk_converter]
 #launch_run clk_converter_synth_1